Allow managed configurations to be saved async
This adds a new #asyncSave()V method to ManagedConfiguration, which serialises the configuration to a string and then writes it to disk in an async thread. This is useful because for large files, saving to disk in the main server thread can cause some considerable lag. This patch also changes ConfigurableMTCModule to use this method by default for #save()V.