[1.6.6] Refactor Module Framework with Dependency Injection
This commit changes the internal architecture of the Module Framework from a mostly-static API based on MTCModuleAdapter to a more dynamic, instance-based approach. In this new architecture, all API access is done through ModuleManager, which is stored in a field of MTC. Internal loading is done via the package-private ModuleLoader. No changes are necessary to existing modules.

Dependency Injection is possible via the new @InjectModule annotation. It injects loaded modules into annotated fields based on their declared type. It offers a parameter to declare a dependency as required, i.e. the module won't load if it isn't fulfilled. Cyclic dependencies are possible if both sides mark the dependency as non-required. Instances are injected upon enable and removed (set to null) when the target module (i.e. the module that is depended upon) is disabled. If the dependency is required, the dependant module is disabled too.

This commit also adds test cases for ModuleLoader. Please treat them with respect, it has been a lot of hassle to get them to work with mocking MTC. Some test-related things have been changed in the poms to allow for the tests to run correctly.