Refactor FullTag distribution to use a retrieval command
Full credit for the idea goes to @Janmm14.

This commit changes the FullTag distribution system. Previously, full items
would be stored in the player's inventory immediately and storage would be
retried in a set interval. This, however, might be a problem to some players
since they might just be in a fight or minigame, meaning they would lose the
full item.

With this commit, they get a notification about the full items waiting for them
when they are created and upon joining the game. To actually retrieve the items,
they have to execute the /canhasfull (/fulls) command. This could also allow
to lift the "no full items for offline players" restriction, but I have not
done this yet since I want to avoid UUID troubles, typos, etc. With online
players, the risk for that seems much lower.