Make /cc send alternating messages; Fixes #451
This apparently helps with some hack clients' frontends combining messages. On the contrary, it is trivial to circumvent this not only by combining not only the most recent message, but also by simply checking the launcher log. But yeah, apparently it shows skiddies that we're professional at preventing hacks or something. If you need valid justification for this commit, yell at Janmm14.