mtc.ignore permission ignores freecam check + range reduce
(Higher) team members occasionally stay in bases somewhere and want to
open an inventory for example with /oe or /oi at it gets annoying
to correct look position for that to work.

Next I reduced the search range of the getLineOfSight method to 8
blocks for performance (and maybe unneeded chunk loads caused by
that).

I also removed the unneccessary player cast and player null check
and now check if the human entity is a player before any
calculation occurs.